refactor: :recycle: Changing the primary key type for searching in repositories

+++ src/main/java/com/thomasdixini/autodex/application/dtos/product/create/InputForProductDto.java
@@ -1,6 +1,6 @@
 package com.thomasdixini.autodex.application.dtos.product.create;
 
 public record InputForProductDto(
-    long inputId,
+    int inputId,
     int quantity
 ) {}
